Move response_mode.go to oauth package
So that they could be reused by the end session handler to rediect back to
the web app after logout
